Fix(models.py): rename user table to account
Since user is a reserved keyword in PostgreSQL, any attempts to create a table
named 'user' will fail. Although it is possible to refer to it with double
quotes ("user"), renaming it seems like the simpler, easier-to-remember option.

All references to instances/attributes of the user table have also been
adjusted accordingly.